Code clean up & Separate runner and model
