+++ constants.py
@@ -31,3 +31,8 @@ STADIUM_NAMES = {
     '고척': 'GoChuk',
     '수원': 'SuWon'
 }
+
+## NN paramters
+LEARNING_RATE = 0.001
+TRAINING_EPOCHS = 500
+KEEP_RATE = 0.7
