--- index.py
@@ -1,4 +1,0 @@
-## NN paramters
-learning_rate = 0.001
-training_epochs = 500
-keep_rate = 0.7
